tests/wpt/web-platform-tests/paint-timing/fcp-only/fcp-text-input.html

<!DOCTYPE html>
<head>
<title>Performance Paint Timing Test: FCP due for non-empty text input</title>
</head>
<body>
<script src="../resources/utils.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<input id="input" type="text" />
<script>
  promise_test(async t => {
      assert_precondition(window.PerformancePaintTiming, "Paint Timing isn't supported.");
      await new Promise(r => window.addEventListener('load', r));
      await assertNoFirstContentfulPaint(t);
      const input = document.getElementById('input');
      input.setAttribute('value', '  ');
      await assertNoFirstContentfulPaint(t);
      input.setAttribute('value', 'value');
      await assertFirstContentfulPaint(t);
  }, 'Text input should become contentful when its value is non-empty');
</script>
</body>
</html>
